Moved remote data folder related common functionality to new file
