+++ crates/modelardb_common/src/lib.rs
@@ -20,3 +20,4 @@ pub mod macros;
 pub mod schemas;
 pub mod types;
 pub mod arguments;
+pub mod remote_data_folder;
